+++ app/(blog)/components/BlogPost.tsx
@@ -1,4 +1,4 @@
-import { Card } from './card'
+import { Card } from '../../components/card'
 
 import { getBlogPosts, sortByDate } from '@/(blog)/utils'
